Move lib/utils/agentconn to lib/sshagent.
